+++ application/controllers/Welcome.php
@@ -4,24 +4,11 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 class Welcome extends MY_Controller {
 
-	/**
-	 * Index Page for this controller.
-	 *
-	 * Maps to the following URL
-	 * 		http://example.com/index.php/welcome
-	 * 	- or -
-	 * 		http://example.com/index.php/welcome/index
-	 * 	- or -
-	 * Since this controller is set as the default controller in
-	 * config/routes.php, it's displayed at http://example.com/
-	 *
-	 * So any other public methods not prefixed with an underscore will
-	 * map to /index.php/welcome/<method_name>
-	 * @see https://codeigniter.com/user_guide/general/urls.html
-	 */
 	public function index() {
-		$this->data['timeslots'] = $this->timeslot_dropdown();
-		$this->data['days'] = $this->days_dropdown();
+		$form = array();
+		$form['days'] = $this->days_dropdown();
+		$form['timeslots'] = $this->timeslot_dropdown();
+		$this->data['theBody'] = $this->parser->parse('_searchForm', $form, true);
 
 		$this->data['pagebody'] = 'welcome';
 		$this->render();