+++ assets/js/js_dataTable.js
@@ -1,0 +1,11 @@
+$(document).ready(function() {
+	$('#courseResult').DataTable({
+		'searching': false
+	});
+	$('#daysResult').DataTable( {
+		'searching': false
+	});
+	$('#timeslotsResult').DataTable({
+		'searching': false
+	});
+} );